[codeReview/knt] Review fix (19)Remove business logic under service
