+++ laravel/app/Contracts/Services/Assignment/AssignmentServiceInterface.php
@@ -96,4 +96,26 @@ interface AssignmentServiceInterface
      * @return $assignemtnList
      */
     public function getAllAssignmentByCourse($course_id);
+
+    /**
+     * To check all assignments for $course_id completed or not
+     * @param $course_id
+     * @return $assignmentStatus
+     */
+    public function isCompletedAssignment($student_id, $course_id);
+
+    /**
+   * check all the assignments are completed or not
+   * @param $student_id, $course_id
+   * @return -> true or false
+   */
+  public function checkAllAssignmentCompleted($student_id, $course_id);
+
+  /**
+     * To show assignment is started or not
+     * @param $course_id
+     * @param $student_id
+     * @return View courseDetails
+     */
+    public function showStarted($student_id, $course_id);
 }
